backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
dj-database-url==2.1.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
pydantic==2.5.3
python-dateutil==2.8.2
psutil==5.9.7
gunicorn==21.2.0
drf-spectacular
